Controls & tooltips z-index.
Since no other element have a z-index defined, 0 is considered for all elements.

Having z-index 50 for the map controls and z-index 1000 for the tooltips can create overlay issues in several cases.

I'm suggesting z-index 1 for map controls, and z-index 2 for tooltips.

This should avoid lots of issues with Boostrap and other CSS framework implementations, as they generally have higher z-index values defined.

A clear example where z-index 50 for map controls is too high, is on a layout that has, let's say... a "slide-over" menu, with a z-index less than that, the map controls would actually overlay that menu, which is really undesirable.

I have already done the required tests to ensure that it still works as expected, tooltips will overalay everything, and the map controls will overlay the map elements.